init M3SDA and test
